backend/requirements.txt
# ==================================================
# FastAPI y servidor
# ==================================================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ==================================================
# Base de datos
# ==================================================
SQLAlchemy>=2.0.25
PyMySQL>=1.1.0

# ==================================================
# Autenticación y seguridad
# ==================================================
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
cryptography>=42.0.0

# ==================================================
# Utilidades
# ==================================================
email-validator>=2.1.0
dnspython>=2.6.0
